Add jackknife function (see #66) in the sandbox module (see #67)
jackknife: repeatedly iterate over seq, each time omitting a successive element

This is the first addition to the sandbox.   The purpose and usage of the
sandbox still needs to be documented.

The goal for `jackknife` is to eventually be in a `sampling` module, which
may end up being a separate project.