`get` is faster for list of indices with no default
PR #57 put an emphasis on the performance of `get` because it is
so frequently used.  Some timings for `get` were given in that thread.
This commit improves those benchmarks as follows:

"test list" is 50% faster (using a two-element list of indices).
"test long list" is 100% faster (using a 100-element list of indices).

This is achieved because `operator.itemgetter(*ind)(seq)` is significantly
faster than `tuple(seq[i] for i in ind)`.  For me, it is 3x faster.