Add keyword argument `numargs` to `curry`
This is useful to enforce variadic arguments to be non-empty, and when
currying extension types so `TypeError` will be properly raised.  I want
this feature for myself, and I will use `numargs` for `cytoolz.curried`
for better error propagation.

Moreover, `numargs` can be used for `curried.merge_with` instead of
redefining a new function in "toolz/curried_exceptions.py".

In this commit, `curry` remains a single class.  I think it would be
cleaner to have `curry` be a function and `Curry` a class.  It's awkward
to handle a keyword argument for `curry`, and it would be much cleaner if
the signature of `Curry` were `Curry(func, args, kwargs, numargs=None)`

I would appreciate feedback on the addition I made to the docstring that
explains the `numargs` keyword.