Test dicttoolz with multiple types of mutable mappings.
This was done by encompassing the tests inside a class, and different
mapping types are tested by subclassing the base class.  Each class
must specify "D", the type of mapping, and "kw", the dict to provide
the "factory" keyword if necessary.  These tests are runnable with both
nosetests and py.test.

`dicttoolz.update_in` was updated so creation of nested mappings use the
factory function, not dict.