ENH: Set __wrapped__ attribute curry and memoize.
Several functions in the stdlib ``functools`` module set this attribute
on decorated functions (most notably, functools.partial).  Consequently,
many introspection tools (e.g. IPython's ? and ?? operators) use
``__wrapped__`` to find the source for functions.

This change makes ``curry`` and ``memoize`` provide a ``__wrapped__``
attribute so that users searching for the source of a curried function
find the actual function source rather than the source for ``curry``.