Much better function introspection for `curry`.
The heavy-lifters are `inspect.getargspec` for Python 2 and
`inspect.signature` for Python 3.  We use these in our new functions
`functoolz.is_valid_args` and `functoolz.is_partial_args`, which
determine via introspection whether a function call is valid or
incomplete with the possibility of becoming valid.

Support for many more builtins was added.  We maintain a registry
of valid signatures in "_signatures.py" for `builtins`, `functools`,
`itertools`, and `operator` modules.

Added `call` and `bind` methods to `curry`.

Removed `_num_required_args`, as it is no longer needed.