Use our introspection registry when `inspect.signature` fails due to inspecting properties.
Unfortunately, `inspect.signature` fails to inspect classes that define
`__signature__` as a property.  Python 3.3 also fails when `__wrapped__`
is a property.  We add these specific edge cases to our inspection handling
and use our registry if possible.  This is needed to support `cytoolz`.

I bet we could fix this in future versions of Python by submitting a
bug/patch to https://bugs.python.com