Use `==` instead of `is` to check for no_default
Previously, `is` was used to check if keywords were equivalent to
`no_default`. This failed if the default string was serialized using
pickle, as they the reconstructed string isn't the same object. This PR
fixes that by replacing all `is` checks for `no_default` with `==`. Some
rearranging of the checks is done to avoid repeating comparisons. Tests
are also added to ensure this behavior remains consistent.

Fixes #309.