Use custom serialization logic via `curry.__reduce__`.  Issue #324.
This allows module-level globals (and their derivatives) created via `@curry`
to be serialized.  For example, `toolz.functoolz.flip` is curried, and the
original curried function is not resolvable by an import path.

Instead of returning a state via `curry.__getstate__` that includes the func
that is no longer resolvable, we return the name of the function if it is
the original (i.e., `toolz.functoolz.flip` will return `"flip"`).  For
derived instances such as `flip(f)` (which creates a new curried instance),
we return a resolvable import path such as `"toolz.functoolz.flip"` instead
of the function, because the function can't be pickled.

More within-code documentation needed.