Make dissoc smarter.
Dissoc was really fast when the keys to remove were relatively few
compared to the size of the dictionary.  This version will pick the smaller
container to iterate over, either building up a new dictionary or tearing
down a copy.

A small heuristic is used to decide if the overhead of building a set is
worth the effort.  My rudimentry testing shows that if the number of keys is
less than 60% the length of the dictionary, tearing down a copy is faster.
If keys is very large, we calculate the intersection between keys and mapping
and build a new dictionary out of that intersection.