remove failing test
Previously I planned to force chunksize value to be larger than 2 and wrote a test to allow chunksize equal to two.
Fold takes a combination function which I am not sure how to interpret and prefer not to mess with so if the user choses a chunksize of 1 than we raise an exception instead of trapping ourselves in an infinite loop.