test_inspect_wrapped_property: handle fixed wrapper inspection
Python upstream recently fixed the behavior of inspect with
wrappers: https://github.com/python/cpython/issues/112006 . The
assertion here relies on the broken behavior, we only get None
if `inspect(Wrapped)` fails and raises `ValueError`. Now it
works, we actually get the correct answer, 1. This changes it so
we assert the correct thing depending on the Python version (the
fix was backported to 3.11.9 and 3.12.3, so the check has to be a
bit complicated).

Signed-off-by: Adam Williamson <awilliam@redhat.com>